Added incrementing hunger based on multiple time increments

+++ src/main/java/com/arkvis/petsim/Pet.java
@@ -13,7 +13,7 @@ public class Pet {
         return name;
     }
 
-    public int getHunger() {
+    public long getHunger() {
         return hunger.getValue();
     }
